moved NDArray types to remove import cycle
